rating dialog setup

+++ app/src/main/java/uz/suhrob/movieinfoapp/data/remote/response/RateBody.kt
@@ -5,5 +5,5 @@ import kotlinx.serialization.Serializable
 
 @Serializable
 data class RateBody(
-    @SerialName("value") val value: Float
+    @SerialName("value") val value: Int
 )
